Add JWT config for RSA and ECDSA
